add language init question, change watch to build:watch
