+++ packages/geoprocessing/scripts/init/createProject.test.ts
@@ -30,6 +30,7 @@ describe("createProject", () => {
         organization: "Test Org",
         repositoryUrl: "https://github.com/test/test-project",
         region: "us-west-1",
+        languages: [],
         templates: [],
         planningAreaType: "other",
         bboxMaxLat: 0,
@@ -78,6 +79,7 @@ describe("createProject", () => {
         organization: "Test Org",
         repositoryUrl: "https://github.com/test/test-project",
         region: "us-west-1",
+        languages: [],
         templates: [],
         planningAreaType: "eez",
         planningAreaId: "Micronesian Exclusive Economic Zone",
@@ -154,6 +156,7 @@ describe("createProject", () => {
         organization: "Test Org",
         repositoryUrl: "https://github.com/test/test-project",
         region: "us-west-1",
+        languages: [],
         templates: ["template-ocean-eez"],
         planningAreaType: "other",
         bboxMaxLat: 0,
@@ -190,6 +193,7 @@ describe("createProject", () => {
         organization: "",
         repositoryUrl: "",
         region: "us-west-1",
+        languages: [],
         templates: [],
         planningAreaType: "other",
         bboxMaxLat: 90,
