+++ packages/geoprocessing/src/i18n/lang/en/translation.json
@@ -4,10 +4,10 @@
   "(Not answered)": "(Not answered)",
   "Activity": "Activity",
   "Area": "Area",
-  "Area Within Plan": "Area Within Plan",
   "Attributes": "Attributes",
   "BlankCard": "BlankCard",
-  "BlankCard Message": "This is a blank report.",
+  "BlankCard Sketch Message": "This is a blank sketch report for geography  with result .",
+  "BlankCard SketchCollection Message": "This is a blank sketch collection report for geography  with result .",
   "Boundary": "Boundary",
   "Category": "Category",
   "Class": "Class",
@@ -63,6 +63,7 @@
   "MPA": "MPA",
   "Map": "Map",
   "Nearshore\n(0-12 nautical miles)": "Nearshore\n(0-12 nautical miles)",
+  "No attributes found": "No attributes found",
   "Offshore\n(12-200 nautical miles)": "Offshore\n(12-200 nautical miles)",
   "OverlapCard": "OverlapCard",
   "OverlapCard - learn more": "<p>ℹ️ Overview:</p><p>🎯 Planning Objective:</p><p>🗺️ Source Data:</p><p>📈 Report: This report calculates the total value of each feature within the plan. This value is divided by the total value of each feature to obtain the % contained within the plan. If the plan includes multiple areas that overlap, the overlap is only counted once.</p>",
