+++ packages/geoprocessing/src/testing/fixtures/projectConfig.ts
@@ -8,6 +8,7 @@ import cloneDeep from "lodash/cloneDeep.js";
 const simpleConfig: ProjectClientConfig = {
   basic: {
     bbox: [0, 0, 1, 1],
+    languages: [],
     planningAreaName: "undefined",
     planningAreaId: "undefined",
     planningAreaType: "eez",
