+++ packages/geoprocessing/src/testing/project/basic.json
@@ -1,5 +1,6 @@
 {
   "bbox": [0, 0, 0, 0],
+  "languages": [],
   "planningAreaType": "eez",
   "planningAreaId": "undefined",
   "planningAreaName": "undefined",
