+++ packages/template-blank-project/project/basic.json
@@ -2,6 +2,7 @@
   "bbox": [
     135.312441837621, -1.17311096529859, 165.676528225997, 13.4454329253893
   ],
+  "languages": [],
   "planningAreaType": "other",
   "planningAreaId": "Micronesia",
   "planningAreaName": "Micronesia",
