+++ packages/template-ocean-eez/project/basic.json
@@ -3,6 +3,7 @@
     -174.51139447157757, -15.878383591829206, -170.54265693017294,
     -10.960825304544073
   ],
+  "languages": [],
   "planningAreaType": "eez",
   "planningAreaId": "Samoan Exclusive Economic Zone",
   "planningAreaName": "Samoa",
